added corrupt savedata detection

+++ scripts/savedata_write.gml
@@ -2,14 +2,16 @@
 
 var b;
 
-if (savedata("backup")) {
-    savedata("backup",false)
-    file_delete(global.backfile)
-    file_copy(global.savefile,global.backfile)
-}
+if (savedata("exists")) {
+    if (savedata("backup")) {
+        savedata("backup",false)
+        file_delete(global.backfile)
+        file_copy(global.savefile,global.backfile)
+    }
 
-b=buffer_create()
-buffer_write_hex(b,ds_map_write(global.savemap))
-if (global.encrypt_save_password!="") buffer_rc4(b,global.encrypt_save_password)
-buffer_save(b,global.savefile)
-buffer_destroy(b)
+    b=buffer_create()
+    buffer_write_hex(b,ds_map_write(global.savemap))
+    if (global.encrypt_save_password!="") buffer_rc4(b,global.encrypt_save_password)
+    buffer_save(b,global.savefile)
+    buffer_destroy(b)
+}
